pyproject.toml
[project]
name = "golden-formulas-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "supabase>=2.27.2",
]
